Created Server Function to store simulated data in JSON File
Server Function to Store Simulated Data in a JSON File : The server will receive the registration details and store the new user's data in a JSON file. If the user already exists, it will prevent duplicate registrations. Here's how you can implement the function on the server side to store the data in a JSON file.